requirements.txt
asgiref==3.6.0
Django==4.1.7
django-tinymce==3.5.0
djangorestframework==3.14.0
Pillow==9.4.0
psycopg2-binary==2.9.5
python-dotenv==1.0.0
pytz==2022.7.1
six==1.16.0
sqlparse==0.4.3
transliterate==1.10.2
